backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.0.1
pyjwt==2.8.0
dbutils==3.0.3
slowapi==0.1.8
python-multipart==0.0.6
email-validator==2.1.0
a2wsgi==1.8.0
sqlalchemy
